Fix Image Orientation Bug in Firefox.
This fix does not help other browsers which currently do not support the
image-orientation CSS property. Coincidentally iOS Safari has the
behavior by default because it automatically respects EXIF data for
orientation.

A good follow-up change for this would be to submit patches to Chromium
that implement this property. This shouldn't be a complete overhaul
given that Chromium will currently respect EXIF orientation data when an
image is viewed in a new tab.