Remove Cruft Associated with Root Directory ID = 3
Because Gallery used to use physically backed files and was located at
`/gallery-data/root/root` Gallery instances had a root directory with an
ID of 3. This created a need to artificially limit viewing of
directories to require a directory ID of at least 3.

Commit 4682ce6fad2f0e5bad4618d80cbe33e4fb8d4101 adds support for S3 and
eliminates the need for the root directory to have a ID of 3 workaround.
This change cleans up that cruft to ease in the setup of future Gallery
instances.

Effort Towards Bug #16